Update tutorials index to provide a roadmap.
Here is my idea for how to make the tutorials index easier to grasp for
new users. This replaces the index.html file in the /tutorials directory to
provide an organized roadmap through all of the existing tutorials,
rather than the by date "blog posting" format we had previously. The
previous index has been moved to all-by-date.html.

Unfortunately this means that new tutorials won't be automatically added
to the tutorial index via jekyll but writing this up by hand
seemed to be the fastest and easiest way to do this.